hsl colour picker refined

+++ Shared/Models/Styling/RadialGradient.cs
@@ -17,6 +17,6 @@ namespace HybridPages.Shared.Models.Styling
 		public int L { get; set; }
 		public float A { get; set; }
 		public int Alpha { get; set; }
-		public int LayerHeight { get; set; }
+		public float LayerHeight { get; set; }
 	}
 }
